sam-app/rss_notification/requirements.txt
requests
feedparser
beautifulsoup4
gspread
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib